:art remoção do index a configurção do redux/saga e colocadas em arquivos especiais
